feat: Map Component Detached
